move POCO to DomainObjects
