+++ v3.0/Source/Kigg.Infrastructure.EF/DomainObjects/CommentSubscribtion.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Kigg.Infrastructure.EF.POCO
+namespace Kigg.Infrastructure.EF.DomainObjects
 {
     public class CommentSubscribtion: Entity
     {
