+++ v3.0/Source/Kigg.Infrastructure.EF/DomainObjects/CommingEvent.cs
@@ -1,4 +1,4 @@
-﻿namespace Kigg.Infrastructure.EF.POCO
+﻿namespace Kigg.Infrastructure.EF.DomainObjects
 {
     public class CommingEvent: Entity
     {
