+++ v3.0/Source/Kigg.Infrastructure.EF/DomainObjects/KnownSource.cs
@@ -1,6 +1,6 @@
 ﻿using Kigg.DomainObjects;
 
-namespace Kigg.Infrastructure.EF.POCO
+namespace Kigg.Infrastructure.EF.DomainObjects
 {
     public class KnownSource: Entity
     {
