+++ v3.0/Source/Kigg.Infrastructure.EF/DomainObjects/StoryComment.cs
@@ -1,7 +1,11 @@
-﻿namespace Kigg.Infrastructure.EF.POCO
+﻿namespace Kigg.Infrastructure.EF.DomainObjects
 {
-    public class StoryMarkAsSpam: Entity
+    public class StoryComment: Entity
     {
+        public string HtmlBody { get; set; }
+
+        public virtual string TextBody { get; set; }
+
         public System.Guid StoryId { get; set; }
         public Story Story { get; set; }
 
@@ -10,6 +14,6 @@
 
         public string IPAddress { get; set; }
 
-        public System.DateTime Timestamp { get; set; }
+        public bool IsOffended { get; set; }
     }
 }
