+++ v3.0/Source/Kigg.Infrastructure.EF/DomainObjects/UserScore.cs
@@ -1,4 +1,4 @@
-﻿namespace Kigg.Infrastructure.EF.POCO
+﻿namespace Kigg.Infrastructure.EF.DomainObjects
 {
     public class UserScore: Entity
     {
@@ -6,7 +6,7 @@
 
         public System.DateTime Timestamp { get; set; }
 
-        public DomainObjects.UserAction ActionType { get; set; }
+        public Kigg.DomainObjects.UserAction ActionType { get; set; }
 
         public decimal Score { get; set; }
