+++ v3.0/Source/Kigg.Infrastructure.EF/DomainObjects/UserTag.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Kigg.Infrastructure.EF.POCO
+namespace Kigg.Infrastructure.EF.DomainObjects
 {
     public class UserTag: Entity
     {
